project/requirements.txt
# Core dependencies for the official workflow and Streamlit judge view.
# Keep this file lean to reduce environment creation time.
#
# System dependency (not installable with pip):
# - A LaTeX distribution providing `pdflatex` for report compilation.
#   Windows: MiKTeX
#   Linux: TeX Live (texlive-latex-base / texlive-full)

# Core data and preprocessing
numpy>=2.0,<3.0
pandas>=2.2,<3.0
scipy>=1.13,<2.0
scikit-learn>=1.5,<2.0
openpyxl>=3.1,<4.0

# App and plotting
plotly>=5.20,<7.0
streamlit>=1.40,<2.0

# Classical models used in the main pipeline
torch>=2.4,<3.0
xgboost>=2.1,<4.0

# Quantum / photonic stack used in the main pipeline
perceval-quandela>=1.1,<2.0
merlinquantum>=0.3,<1.0
